Use the 'develop' branch when editing a post on GH
When clicking on 'Edit this post on GitHub', the user was invited to
edit the post from the 'master' branch which was not possible until the
next release version due to the GitFlow methodology. A 404 error was
returned by GitHub because the file was missing from 'master'.

The user is now invited to edit the post from the 'develop' branch.